+++ src/commonMain/kotlin/baaahs/Pinky.kt
@@ -16,8 +16,11 @@ import baaahs.model.Model
 import baaahs.net.FragmentingUdpLink
 import baaahs.net.Network
 import baaahs.plugin.Plugins
+import baaahs.plugin.beatlink.BeatData
+import baaahs.plugin.beatlink.BeatSource
 import baaahs.proto.*
 import baaahs.show.Show
+import baaahs.util.Clock
 import baaahs.util.Framerate
 import baaahs.util.Logger
 import kotlinx.coroutines.*
