+++ src/commonMain/kotlin/baaahs/util/Time.kt
@@ -1,0 +1,7 @@
+package baaahs.util
+
+typealias Time = Double
+
+interface Clock {
+    fun now(): Time
+}
