+++ src/jsMain/kotlin/baaahs/SheepSimulator.kt
@@ -10,11 +10,13 @@ import baaahs.mapper.MappingSession.SurfaceData.PixelData
 import baaahs.mapper.Storage
 import baaahs.model.Model
 import baaahs.net.Network
-import baaahs.plugin.BeatLinkPlugin
 import baaahs.plugin.Plugins
+import baaahs.plugin.beatlink.BeatLinkPlugin
 import baaahs.proto.Ports
 import baaahs.shows.BakedInShaders
 import baaahs.sim.*
+import baaahs.util.Clock
+import baaahs.util.Time
 import baaahs.visualizer.SurfaceGeometry
 import baaahs.visualizer.SwirlyPixelArranger
 import baaahs.visualizer.Visualizer
