+++ src/jsMain/kotlin/baaahs/app/ui/AppContext.kt
@@ -1,12 +1,12 @@
 package baaahs.app.ui
 
-import baaahs.Clock
 import baaahs.ShowPlayer
 import baaahs.client.WebClient
 import baaahs.gl.patch.AutoWirer
 import baaahs.plugin.Plugins
 import baaahs.ui.Prompt
 import baaahs.ui.ReactBeautifulDragNDrop
+import baaahs.util.Clock
 import react.createContext
 
 val appContext = createContext<AppContext>()
