+++ src/jsMain/kotlin/baaahs/app/ui/editor/TextEditor.kt
@@ -7,12 +7,12 @@ import acex.AceEditor
 import acex.Mode
 import acex.Theme
 import acex.Themes
-import baaahs.Time
 import baaahs.app.ui.appContext
 import baaahs.jsx.useResizeListener
 import baaahs.ui.Styles
 import baaahs.ui.unaryPlus
 import baaahs.ui.xComponent
+import baaahs.util.Time
 import baaahs.window
 import kotlinext.js.jsObject
 import materialui.styles.palette.PaletteType
