+++ src/jvmMain/kotlin/baaahs/SystemClock.kt
@@ -1,5 +1,8 @@
 package baaahs
 
+import baaahs.util.Clock
+import baaahs.util.Time
+
 class SystemClock : Clock {
     override fun now(): Time {
         return System.currentTimeMillis().toDouble() / 1000.0
